+++ packages/angular/src/index.ts
@@ -14,6 +14,7 @@ export {
 	CirclePackChartComponent,
 	ComboChartComponent,
 	DonutChartComponent,
+	ExperimentalChoroplethChartComponent,
 	GaugeChartComponent,
 	GroupedBarChartComponent,
 	HeatmapChartComponent,