+++ packages/angular/src/lib/index.ts
@@ -9,6 +9,7 @@ export {
 	CirclePackChartComponent,
 	ComboChartComponent,
 	DonutChartComponent,
+	ExperimentalChoroplethChartComponent,
 	GaugeChartComponent,
 	GroupedBarChartComponent,
 	HeatmapChartComponent,