+++ packages/core/src/charts/index.ts
@@ -3,6 +3,7 @@ import { AreaChart } from './area'
 import { BoxplotChart } from './boxplot'
 import { BubbleChart } from './bubble'
 import { BulletChart } from './bullet'
+import { ExperimentalChoroplethChart } from './choropleth'
 import { CirclePackChart } from './circle-pack'
 import { ComboChart } from './combo'
 import { DonutChart } from './donut'
@@ -29,6 +30,7 @@ export {
 	BoxplotChart,
 	BubbleChart,
 	BulletChart,
+	ExperimentalChoroplethChart,
 	CirclePackChart,
 	ComboChart,
 	DonutChart,
@@ -56,6 +58,7 @@ export type Charts =
 	| BoxplotChart
 	| BubbleChart
 	| BulletChart
+	| ExperimentalChoroplethChart
 	| CirclePackChart
 	| ComboChart
 	| DonutChart